fix: remove unused code
